Update instructions to work around SSL troubles
Given that GitHub has dropped support for SSL/TLS ciphers that Mac OS X 10.4 and 10.5's built-in `curl` support, the installation instructions are currently broken.

This updates the instructions to instead suggest downloading the install script with their modern computer and transferring it via the same medium they use for Xcode, which hopefully avoids some of the potential pitfalls people have experienced with TenFourFox.