dovecot 2.3.20
Build with PAM support by default, Apple's build of dovecot would patch in
OpenDirectory support so auth would likely work without PAM, however without
such patch auth fails since user account live in OD, not in the passwd file.

Dovecot detects that kqueue is available and tries to use it for ioloop and notify,
however, the version of kqueue in Leopard and Tiger is too old. So while it'll
compile ok, it'll fail at runtime when it tries to set the events on startup.