clisp build fix
Backport fixes from git repo to fix issue with iconv macro and the
lack of HTTPS support in clisp. The test suite would previously
attempt to connect to the clisp website but since an HTTP->HTTPS
redirect is in place now, it would trip up & the test suite run would
fail.
Add fix to report an error when attempting to connect to HTTPS urls
and fix the test suite to not make network connections.