Tidy up lighttpd formula
Based on feedback in PR #917
Drop attr support
Building with libev is no longer a supported option
Use kqueue support as it's the optimum means
sendfile support is officially supported in 10.5 but is actually
there in 10.4 but guarded off, one of the parameters is a different
type though. Play it safe.
bzip2 encoding was non-standard so skip using it.
https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Accept-Encoding