qemu 2.3.1
Previously, this formula was reverted from 2.4.x to 2.2.1 due to
lockups experienced on 10.5.8 on G5. 2.3.1 runs ok for me on a G5
running Tiger.
Drop gnutls as it is dangerously out of date and disable VNC TLS support
which requires it.
Allow the guest agent to be built in case OS X is virtualised.
The user mode is not enabled by default, there's no need to disable it.
Update the url for the test image with a working URL.

Built on Tiger with GCC 5
i386 hypervisor tested with
qemu-system-i386 -cdrom some-bootable.iso
PowerPC hypervisor tested with
qemu-system-ppc some.iso
Then connected to the console via VNC on localhost using Chicken.