openssl: Tiger/Intel build fixes
Ensure makedepend is never used as it slows down the build considerably.

For PowerPC builds OpenSSL already applies -fno-common automatically
but opted to pass it conditionally here to avoid having to patch OpenSSL.

as(1) for Tiger/Intel only recognises ".comm name, size", not ".com name, size, alignment"
We're also building with no-common and the assembly files use the common directive. :/