libiconv 1.18
Use the stock Makefile rather than Makefile.devel as the latter has
an autotools dependency and I suspect it was used history because
it respected CC & CFLAGS. It turns out so does Makefile now.
Tested by building with --cc stated & without as well as well as
ENV.no_optimization set to ensure that set CC & CFLAGS values are
made use of.
Run the test suite for peace of mind, it is not a lengthy run.