python3: Switch to using libffi formula
With the upgrade to 3.10.16, builds on a Tiger G4 resulted in the
ctype module not building. It turns out that it libffi was being
linked against opportunistically via pkg-config in some scenarios.
Just depends on libffi since the version found in the source of
Python (Modules/_ctypes/libffi_osx) is very old (a pre-release of 2.0).

Some modules also link against zlib so depend on that too.
Add a test to ensure the ctype module works in that it imports as
the build of module can fail and the build of Python can succeed only
to find that there no module to import later.
While here add a guard for tkinter test since only on Xorg X11 in 10.5
and up will X11 be started and the test run within. On Tiger you have
in invoke the test from xterm otherwise it'll just fail.