flac 1.4.3
Drop universal build support as build was broken on new version.
--enable-sse no longer recognised.
configure sets things up, no need to set OBJ_FORMAT, assembly support
is enabled if building with clang.
Due to shortcoming in clang shipped with 10.8/10.9, llvm-gcc needs to
be used, but there's no longer an issue building with llvm-gcc.
Enable OGG support, build times are for both are very short.
Tested on 10.4 to 10.11.
Resolves #335