gcc48: revisit patches
Given an up to date copy of El Capitan, only the patch from the
GCC bug is required.
The macros issue originated when El Capitan was still in beta and
was addressed by Apple in XCode 7.3.
Skip the conditional as the fix from the GCC bug was integrated into
GCC.