gcc47: tweak build
--enable-libstdcxx-time alters the C++11 ABI and shouldn't be used. https://gcc.gnu.org/bugzilla/show_bug.cgi?id=61758#c6
Allow the build itself to set the type of checking.
We're not building head so no need to specify disable werror.
Update URLs
Don't blindly enable LTO, it is not supported on PowerPC https://github.com/iains/darwin-toolchains-start-here/discussions/37#discussioncomment-3285847
Add a test for C++ & Fortran
Skip the conditional as the fix from the GCC bug was integrated into
GCC.