gcc44: tweak build
Start to make this formula look like the other GCC formulas.
Turn off optimisation as build breaks
--enable-version-specific-runtime-libs doesn't work here, so follow
the same procedure for guiding the build for linking against libraries
otherwise the build succeeds but comes tumbling down when g++ is invoked
as libgcc_s cannot be found.
--enable-libstdcxx-time alters the C++11 ABI and shouldn't be used. https://gcc.gnu.org/bugzilla/show_bug.cgi?id=61758#c6
Allow the build itself to set the type of checking.
We're not building head so no need to specify disable werror.
Update URLs