openssl3: Use usleep when on G5 running 10.5
So the test suite run passes.
OpenSSL 3.4.0 switched to using nanosleep(3) by default ond when
running Leopard on a G5, the sleep sanity test which calls nanosleep
fails the test. As a workaround, switch to usleep.
Issue is not observed on a G4 running Leopard, nor the same G5 running
10.4.
https://github.com/openssl/openssl/issues/27317