syntaxerl 0.13.0
Update syntaxerl to version 0.13.0, later versions do not produce a `syntaxerl` program in /usr/local/bin/ which seems to be due to a `rebar3` dependency, which itself depends on OTP 19. No release notes are provided in Github for syntaxerl, so this may be resolvable under OTP 18, but will require more investigation.