ejabberd: 16.12
Update ejabberd to 16.12. Version 16 is the first version I was able to build on Leopard/PPC, and 16.12 is the highest minor version of ejabberd 16.

Also switch to using openssl3.

Tested on Leopard/PPC. I would love help testing on Snow Leopard if you're able to, as the pam_appl.h workaround I needed to add is apparently not needed on SL, but I don't currently have a way to test that.