erlang: Fix build on Snow Leopard
erlc built with GCC 4.2 fails with Bus error when it reaches HiPE build.
Switch to llvm-gcc instead.
llvm-gcc may have had issues compiling OTP 14 & 16, but it work here for 18.
Originally the exclusion was for 14 for llvm-gcc bulid 2334, then
changed to an outright exclusion for OTP 16. Seems to be fine on
10.6 with llvm-gcc build 2335.6.