erlang: allow current autoconf to work
We sidestepped dealing with this when the formula was upgraded to v18
by making use of the autoconf shipped with OS X which is fine up
until 10.7 where autoconf was removed.
Tested on 10.7 to 10.11.
Remaining issue is on 10.11 erlang builds but will not link to the
version of OpenSSL included with OS, need to fix crypto.c so OpenSSL 3
can be used.