scheme48 1.9.3
Source archive comes with documentation already generated from tex files.
Use them instead of regenerating which would pull in a LaTeX dependency.
No need to adjust optimisation level when building with clang (tested on 10.7 to 10.11)
bibop garbage collector is the default.
Add an option to run the test suite.

This is the first formula where I've seen noticable difference in performance
with the compiler optimisation. With default optimisation settings it took
17 minutes to build & run the test suite. With no_optimization set, it took
55 minutes to build & run the test suite. Tested on a 1.8Ghz iMac G5.