libaacs 0.9.0
v0.9.0 is the most recent version which compiles out of the box on Leopard.
It's possble to get this to compile on Tiger but I decided on not including
the patch since the most recent version of VLC which runs on Tiger is v0.9.10.
I also lack the hardware and any other software which could use libaacs in order
to be able to test beyond compilation. Let me know if you want to give it a try.