git: Fix build on Leopard
To build Git-Gui.app Tcl-Tk 8.5 with Aqua support is required.
Snow Leopard shipped with Tcl-Tk 8.5. Unable to use an up to date
version of Tcl-Tk via Tigerbrew since Tk requires functionality in
AppKit introduced in Snow Leopard for Aqua support.
Perhaps an option is to not build the .app and see if it can be run
via X11 support on Leopard and prior.

Fixes issue #1410