tcl-tk: Tk build fixes
GCC 4.2 on 10.6 doesn't like the turnary operation there, but if you
split it out, it works. Fixes #1361
Add missing brace to allow the test suite to run on 10.7 & prior.

Can't use the usual patch method which Tigerbrew provides to patch
a resource, so have to invoke patch(1) via system method.