Only run test suite if requested
or when bottling.
The road to building Mercurial & git on a fresh setup should be a
little quicker now. On a stock OS it should be expected to work without
needing to run the test suite every time, as things should've been
tested when preparing to commit the update to Tigerbrew.