links 2.30
Rework options, hang image support off X11 support, and make X11 support default.
Pull in up to date deps, with the exception of fontconfig and by indirect dependency
freetype as it causes issues starting in graphics mode on 10.6, and
has a build issue on 10.7 which need to be resolved before they can
be depended on.