screen: Add formula for v4.9.1
The v4 branch is C99 based, with v5 it switched to C17.
The legacy v4 version which Apple still ships with sets TIOCSCTTY
if it is defined, I have followed the same here rather than guarding
it off to Tiger, since there is no error on newer releases without
setting TIOCSCTTY.
It's not possible to control whether utmp is used during configure so
have to resort to the old fashioned means of handling it in config.h.
Skipped on the use of utmp since it warns about deprication during build
and not finding the utmp file during runtime on Leopard and newer.

Have test that it's happy to start, create a new window, alternate,
detach, re-attach, and use with a serial port.
Tested with a Prolific PL2303 USB to RS-232 adapter and some
network hardware at 9600 baud on OS X Tiger and Leopard.

Adapted from upstream
https://github.com/Homebrew/homebrew-core/blob/109168dd36e60dcb9e2609929281356d6d53bb26/Formula/s/screen.rb