stlink 1.7.0
This is the last version which support macOS. v1.8.0 drops support.
To build with GCC 4.0 on Tiger, need to enable warnings as passing
the flag to turn them off to libtool trips it up, but if you enable
warnings, can't treat them as hard errors.
GCC 4.0 defaults to C89.

Tested that st-info was able to probe an st-link v2 module, and to
use st-util to spawn a gdbserver which arm-none-eabi-gdb was able to
attach to where an stm32-discovery board was connected to the st-link v2
module for SWD.