docs: use vendored ruby & modern curl
curl included with the OS most definitely will be unable to fetch
the script due to ancient OpenSSL, and the uninstall script needs
a newer version of ruby than is provided at the very least in
Leopard and Tiger.

uninstall script uses String#start_with?
String#start_with? showed up in Ruby 1.8.7.
Ruby shipped in Leopard is 1.8.6

Discovered in issue #1459