guile 3.0.11
Signature for iconv on Tiger is different
Provide a prebuilt set of images for PowerPC, built on a G5 running Tiger
but basic testing looked fine on a G4 running Leopard.
Guile built with Tigerbrew's default optimisation setting frequently segfaulted.
Work in progress, needs tidying e.g the moving of stage directories can be one liner
in a loop for each part.